Fix issue where mouseMoved event fires with previous screen
Also create a global variable for the eventtap so that PaperWM can
disable it when Mission Control is open. This mouseMoved eventtap
prevents events from reaching Mission Control, even when we explicitly
propagate the received events.

I was also unable to detect when Mission Control is open with a
uielement observer.